refactor: JWT 패키지 구조 변경 및 oauth 코드 리팩토링
