+++ backend/src/main/java/com/ll/readycode/global/common/auth/oauth/service/KakaoOAuthService.java
@@ -1,7 +1,7 @@
 package com.ll.readycode.global.common.auth.oauth.service;
 
 import com.ll.readycode.domain.users.userauths.repository.UserAuthRepository;
-import com.ll.readycode.global.common.auth.jwt.provider.JwtProvider;
+import com.ll.readycode.global.common.auth.jwt.JwtProvider;
 import com.ll.readycode.global.common.auth.oauth.dto.KakaoTokenResponse;
 import com.ll.readycode.global.common.auth.oauth.dto.KakaoUserInfo;
 import com.ll.readycode.global.common.auth.oauth.properties.OAuthProperties;
